+++ web/src/app/router.tsx
@@ -19,6 +19,7 @@ import { RequestDetail } from '../pages/RequestDetail';
 import { People } from '../pages/People';
 import { Contacts } from '../pages/Contacts';
 import { AccountStatement } from '../pages/AccountStatement';
+import { Reconcile } from '../pages/Reconcile';
 import { Businesses } from '../pages/Businesses';
 import { WhoDidWhat } from '../pages/WhoDidWhat';
 import { Notifications } from '../pages/Notifications';
@@ -68,6 +69,7 @@ function Gate() {
         <Route path="/contacts" element={<Contacts />} />
         <Route path="/businesses" element={<Businesses />} />
         <Route path="/accounts/:id" element={<AccountStatement />} />
+        <Route path="/reconcile" element={<Reconcile />} />
         <Route path="/who-did-what" element={<WhoDidWhat />} />
         <Route path="/notifications" element={<Notifications />} />
         <Route path="/not-possible" element={<NotPossible />} />